requirements.txt
asgiref==3.10.0
dj-database-url==3.0.1
Django==5.2.7
gunicorn==23.0.0
packaging==25.0
pillow==12.0.0
psycopg2==2.9.11
psycopg2-binary==2.9.10
python-decouple==3.8
python-dotenv==1.2.1
sqlparse==0.5.3
typing_extensions==4.15.0
whitenoise==6.6.0
